now it can run on cuda. some differences in results between omp and cuda program, need detailed compare

+++ julia/FMI_IO.jl
@@ -3,12 +3,12 @@ module FMI_IO
 import Base: read, write
 
 struct Trace
-    n_samples::Int64
-    ref_i::Int64
-    pwin_i::Int64
-    pwin_l::Int64
-    swin_i::Int64
-    swin_l::Int64
+    n_samples::Int32
+    ref_i::Int32
+    pwin_i::Int32
+    pwin_l::Int32
+    swin_i::Int32
+    swin_l::Int32
     obs::Vector{Float32}
 end
 
@@ -16,17 +16,17 @@ function Trace(n_samples::Integer, ref_i::Integer,
     pwin_i::Integer, pwin_l::Integer,
     swin_i::Integer, swin_l::Integer,
     obs::AbstractVector{<:AbstractFloat})
-    return Trace(Int64(n_samples), Int64(ref_i), Int64(pwin_i), Int64(pwin_l),
-        Int64(swin_i), Int64(swin_l), Float32.(obs))
+    return Trace(Int32(n_samples), Int32(ref_i), Int32(pwin_i), Int32(pwin_l),
+        Int32(swin_i), Int32(swin_l), Float32.(obs))
 end
 
 function read(io::IO, ::Type{Trace})
-    n_samples = read(io, Int64)
-    ref_i = read(io, Int64)
-    pwin_i = Base.read(io, Int64)
-    pwin_l = read(io, Int64)
-    swin_i = read(io, Int64)
-    swin_l = read(io, Int64)
+    n_samples = read(io, Int32)
+    ref_i = read(io, Int32)
+    pwin_i = Base.read(io, Int32)
+    pwin_l = read(io, Int32)
+    swin_i = read(io, Int32)
+    swin_l = read(io, Int32)
     obs = zeros(Float32, n_samples)
     read!(io, obs)
     return Trace(n_samples, ref_i, pwin_i, pwin_l, swin_i, swin_l, obs)
@@ -78,9 +78,9 @@ function write(io::IO, s::MomentTensor)
 end
 
 struct GFtrace
-    n_samples::Int64
-    p_i::Int64
-    s_i::Int64
+    n_samples::Int32
+    p_i::Int32
+    s_i::Int32
     g11::Vector{Float32}
     g22::Vector{Float32}
     g33::Vector{Float32}
@@ -97,14 +97,14 @@ function GFtrace(n_samples::Integer, p_i::Integer, s_i::Integer,
     g13::AbstractVector{<:AbstractFloat},
     g23::AbstractVector{<:AbstractFloat}
 )
-    return GFtrace(Int64(n_samples), Int64(p_i), Int64(s_i),
+    return GFtrace(Int32(n_samples), Int32(p_i), Int32(s_i),
         Float32.(g11), Float32.(g22), Float32.(g33), Float32.(g12), Float32.(g13), Float32.(g23))
 end
 
 function read(io::IO, ::Type{GFtrace})
-    n = read(io, Int64)
-    p_i = read(io, Int64)
-    s_i = read(io, Int64)
+    n = read(io, Int32)
+    p_i = read(io, Int32)
+    s_i = read(io, Int32)
     g11 = zeros(Float32, n)
     g22 = zeros(Float32, n)
     g33 = zeros(Float32, n)
@@ -133,11 +133,11 @@ function write(io::IO, gf::GFtrace)
 end
 
 function read_c_input(io::IO)
-    ntrs = read(io, Int64)
+    ntrs = read(io, Int32)
     @info "ntrs: $ntrs"
-    nmts = read(io, Int64)
+    nmts = read(io, Int32)
     @info "nmts: $nmts"
-    nsrc = read(io, Int64)
+    nsrc = read(io, Int32)
     @info "nsrc: $nsrc"
     ngfs = ntrs * nsrc
     traces = Vector{Trace}(undef, ntrs)
@@ -159,9 +159,9 @@ end
 read_c_input(fname::AbstractString) = open(read_c_input, fname, "r")
 
 function write_c_input(io::IO, traces::AbstractVector{Trace}, mts::AbstractVector{MomentTensor}, gf::AbstractVector{GFtrace})
-    write(io, Int64(length(traces)))
-    write(io, Int64(length(mts)))
-    write(io, Int64(length(gf)/length(traces)))
+    write(io, Int32(length(traces)))
+    write(io, Int32(length(mts)))
+    write(io, Int32(length(gf)/length(traces)))
     for i = eachindex(traces)
         write(io, traces[i])
     end
@@ -178,13 +178,13 @@ write_c_input(fname::AbstractString, traces::AbstractVector{Trace}, mts::Abstrac
     open(io -> write_c_input(io, traces, mts, gf), fname, "w")
 
 function read_c_result(io::IO)
-    ntrs = read(io, Int64)
-    nmts = read(io, Int64)
-    nsrc = read(io, Int64)
+    ntrs = read(io, Int32)
+    nmts = read(io, Int32)
+    nsrc = read(io, Int32)
     misfit_pl2 = zeros(Float32, ntrs, nmts, nsrc)
-    misfit_pshift = zeros(Int64, ntrs, nmts, nsrc)
+    misfit_pshift = zeros(Int32, ntrs, nmts, nsrc)
     misfit_sl2 = zeros(Float32, ntrs, nmts, nsrc)
-    misfit_sshift = zeros(Int64, ntrs, nmts, nsrc)
+    misfit_sshift = zeros(Int32, ntrs, nmts, nsrc)
     misfit_pol = zeros(Float32, ntrs, nmts, nsrc)
     misfit_psr = zeros(Float32, ntrs, nmts, nsrc)
     read!(io, misfit_pl2)
